requirements.txt
pillow
Flask
tensorflow>=2.0.0